Type fix RV table
In the NEID header order by order RVs, the edge orders without wavelength solutions have their RV values set to None rather than 0 (orders that are processed but not used in the RV calculation). This changes those values to nan when read into the standardized RV1 table so that all entries are the same type.