NEID missing wavelength to nan
In NEID data, there are 6 orders that do not have a wavelength solution (the first 3 and the last 3). In the native data format, the wavelength arrays for these orders are set to all 0s. This commit has the translator change those order entries in the wavelength solution array to all nans.